+++ swing/common-model/src/test/java/is/codion/swing/common/model/component/table/DefaultFilterTableModelTest.java
@@ -131,14 +131,14 @@ public final class DefaultFilterTableModelTest {
 	}
 
 	@Test
-	void filterModel() {
+	void conditionModel() {
 		tableModel.refresh();
 		assertEquals(5, tableModel.items().visible().count());
-		tableModel.filterModel().conditionModel(0).operands().equal().set("a");
+		tableModel.conditionModel().conditionModel(0).operands().equal().set("a");
 		assertEquals(1, tableModel.items().visible().count());
-		tableModel.filterModel().conditionModel(0).operands().equal().set("b");
+		tableModel.conditionModel().conditionModel(0).operands().equal().set("b");
 		assertEquals(1, tableModel.items().visible().count());
-		tableModel.filterModel().conditionModel(0).clear();
+		tableModel.conditionModel().conditionModel(0).clear();
 	}
 
 	@Test
@@ -248,7 +248,7 @@ public final class DefaultFilterTableModelTest {
 		tableModel.items().visible().addListener(listener);
 		tableModel.refresh();
 		assertEquals(1, events.get());
-		tableModel.filterModel().conditionModel(0).operands().equal().set("a");
+		tableModel.conditionModel().conditionModel(0).operands().equal().set("a");
 		tableModel.removeItem(B);
 		assertEquals(3, events.get());
 		assertFalse(tableModel.items().visible().contains(B));
@@ -262,7 +262,7 @@ public final class DefaultFilterTableModelTest {
 		assertFalse(tableModel.items().visible().contains(E));
 		assertFalse(tableModel.items().filtered().contains(D));
 		assertFalse(tableModel.items().filtered().contains(E));
-		tableModel.filterModel().conditionModel(0).operands().equal().set(null);
+		tableModel.conditionModel().conditionModel(0).operands().equal().set(null);
 		tableModel.refresh();//two events, clear and add
 		assertEquals(8, events.get());
 		tableModel.removeItems(0, 2);
@@ -529,10 +529,10 @@ public final class DefaultFilterTableModelTest {
 		tableModel.selection().indexes().add(singletonList(3));
 		assertEquals(3, tableModel.selection().getMinSelectionIndex());
 
-		tableModel.filterModel().conditionModel(0).operands().equal().set("d");
+		tableModel.conditionModel().conditionModel(0).operands().equal().set("d");
 		assertEquals(0, tableModel.selection().getMinSelectionIndex());
 		assertEquals(singletonList(0), tableModel.selection().indexes().get());
-		tableModel.filterModel().conditionModel(0).enabled().set(false);
+		tableModel.conditionModel().conditionModel(0).enabled().set(false);
 		assertEquals(0, tableModel.selection().getMinSelectionIndex());
 		assertEquals(ITEMS.get(3), tableModel.selection().item().get());
 	}
@@ -552,7 +552,7 @@ public final class DefaultFilterTableModelTest {
 	@Test
 	void filterAndRemove() {
 		tableModel.refresh();
-		tableModel.filterModel().conditionModel(0).operands().equal().set("a");
+		tableModel.conditionModel().conditionModel(0).operands().equal().set("a");
 		assertTrue(tableModel.items().contains(B));
 		tableModel.removeItem(B);
 		assertFalse(tableModel.items().contains(B));
@@ -567,9 +567,9 @@ public final class DefaultFilterTableModelTest {
 		assertTrue(tableModel.items().visiblePredicate().isNull());
 
 		//test filters
-		assertNotNull(tableModel.filterModel().conditionModel(0));
+		assertNotNull(tableModel.conditionModel().conditionModel(0));
 		assertTrue(tableModel.items().visible().contains(B));
-		tableModel.filterModel().conditionModel(0).operands().equal().set("a");
+		tableModel.conditionModel().conditionModel(0).operands().equal().set("a");
 		assertTrue(tableModel.items().visible().contains(A));
 		assertFalse(tableModel.items().visible().contains(B));
 		assertTrue(tableModel.items().filtered().contains(D));
@@ -582,7 +582,7 @@ public final class DefaultFilterTableModelTest {
 
 		assertFalse(tableModel.items().visible().contains(B));
 		assertTrue(tableModel.items().contains(B));
-		assertTrue(tableModel.filterModel().conditionModel(0).enabled().get());
+		assertTrue(tableModel.conditionModel().conditionModel(0).enabled().get());
 		assertEquals(4, tableModel.items().filtered().count());
 		assertFalse(tableModelContainsAll(ITEMS, false, tableModel));
 		assertTrue(tableModelContainsAll(ITEMS, true, tableModel));
@@ -591,35 +591,35 @@ public final class DefaultFilterTableModelTest {
 		assertFalse(tableModel.items().filtered().get().isEmpty());
 		assertFalse(tableModel.items().get().isEmpty());
 
-		tableModel.filterModel().conditionModel(0).enabled().set(false);
-		assertFalse(tableModel.filterModel().conditionModel(0).enabled().get());
+		tableModel.conditionModel().conditionModel(0).enabled().set(false);
+		assertFalse(tableModel.conditionModel().conditionModel(0).enabled().get());
 
 		assertTrue(tableModelContainsAll(ITEMS, false, tableModel));
 
-		tableModel.filterModel().conditionModel(0).operands().equal().set("t"); // ekki til
-		assertTrue(tableModel.filterModel().conditionModel(0).enabled().get());
+		tableModel.conditionModel().conditionModel(0).operands().equal().set("t"); // ekki til
+		assertTrue(tableModel.conditionModel().conditionModel(0).enabled().get());
 		assertEquals(5, tableModel.items().filtered().count());
 		assertFalse(tableModelContainsAll(ITEMS, false, tableModel));
 		assertTrue(tableModelContainsAll(ITEMS, true, tableModel));
-		tableModel.filterModel().conditionModel(0).enabled().set(false);
+		tableModel.conditionModel().conditionModel(0).enabled().set(false);
 		assertTrue(tableModelContainsAll(ITEMS, false, tableModel));
-		assertFalse(tableModel.filterModel().conditionModel(0).enabled().get());
+		assertFalse(tableModel.conditionModel().conditionModel(0).enabled().get());
 
-		tableModel.filterModel().conditionModel(0).operands().equal().set("b");
+		tableModel.conditionModel().conditionModel(0).operands().equal().set("b");
 		int rowCount = tableModel.items().visible().count();
 		tableModel.addItemsAt(0, singletonList(new TestRow("x")));
 		assertEquals(rowCount, tableModel.items().visible().count());
 
-		assertThrows(IllegalArgumentException.class, () -> tableModel.filterModel().conditionModel(1));
+		assertThrows(IllegalArgumentException.class, () -> tableModel.conditionModel().conditionModel(1));
 	}
 
 	@Test
 	void clearFilterModels() {
-		assertFalse(tableModel.filterModel().enabled(0).get());
-		tableModel.filterModel().conditionModel(0).operands().equal().set("SCOTT");
-		assertTrue(tableModel.filterModel().enabled(0).get());
-		tableModel.filterModel().clear();
-		assertFalse(tableModel.filterModel().enabled(0).get());
+		assertFalse(tableModel.conditionModel().enabled(0).get());
+		tableModel.conditionModel().conditionModel(0).operands().equal().set("SCOTT");
+		assertTrue(tableModel.conditionModel().enabled(0).get());
+		tableModel.conditionModel().clear();
+		assertFalse(tableModel.conditionModel().enabled(0).get());
 	}
 
 	@Test